requirement.txt
argon2-cffi==19.1.0
autopep8==1.4.4
cffi==1.12.3
DateTime==4.3
Django==2.2.4
djangorestframework==3.10.2
docutils==0.15.2
mysqlclient==1.4.2.post1
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
zope.interface==4.6.0
